feat(opal): move CreateButton into Opal library
Moves CreateButton from refresh-components into lib/opal/src/components/buttons/create-button/.
Migrates all 10 callers to import from @opal/components. Removes rightIcon prop.
